Unrelated method implements IReadOnly interface
Because default interface methods are considered lower priority than the "old" interface resolution algorithm, a suitable public virtual method with the same name/sig will implement the interface method.

If this diff is applied, Application will throw a `NotImplementedException` because the default implementation was ignored in favor of a pre-existing (but unrelated) name/sig match.